+++ js/const.js
@@ -1,0 +1,15 @@
+const dataConst = {
+  TYPE: ['palace', 'flat', 'house', 'bungalow', 'hotel'],
+
+  CHECKS_TIME: ['12:00', '13:00', '14:00'],
+
+  FEATURES: ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'],
+
+  PHOTOS: [
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+  ]
+};
+
+export {dataConst};
